intento de exposicion de api
