+++ pom.xml
@@ -49,6 +49,11 @@
 			<version>1.5.0</version>
 		</dependency>
 		<dependency>
+			<groupId>co.com.sofka</groupId>
+			<artifactId>infrastructure</artifactId>
+			<version>1.5.0</version>
+		</dependency>
+		<dependency>
 			<groupId>junit</groupId>
 			<artifactId>junit</artifactId>
 			<version>4.11</version>
