+++ src/main/java/co/com/sofkau/horseraces/usecases/game/commands/AddTrackUseCase.java
@@ -1,21 +1,21 @@
-package co.com.sofkau.horseraces.usecases;
+package co.com.sofkau.horseraces.usecases.game.commands;
 
 import co.com.sofka.business.generic.BusinessException;
 import co.com.sofka.business.generic.UseCase;
 import co.com.sofka.business.support.RequestCommand;
 import co.com.sofka.business.support.ResponseEvents;
 import co.com.sofkau.horseraces.domain.game.Game;
-import co.com.sofkau.horseraces.domain.game.commands.CreateHorse;
-import co.com.sofkau.horseraces.domain.game.commands.CreatePlayer;
+import co.com.sofkau.horseraces.domain.game.commands.AddTrack;
 
-public class CreatePlayerUseCase extends UseCase<RequestCommand<CreatePlayer>, ResponseEvents> {
+public class AddTrackUseCase extends UseCase<RequestCommand<AddTrack>, ResponseEvents> {
+
     @Override
-    public void executeUseCase(RequestCommand<CreatePlayer> createPlayerRequestCommand) {
-        var command = createPlayerRequestCommand.getCommand();
+    public void executeUseCase(RequestCommand<AddTrack> addTrackRequestCommand) {
+        var command = addTrackRequestCommand.getCommand();
         var game = Game.from(command.getGameId(), retrieveEvents(command.getGameId().value()));
 
-        if (game.getActualState().value().equals("FINISHED")) {
-            game.createPlayer(command.getPlayerId(),command.getName());
+        if (game.getActualState().value().equals("IDLE")) {
+            game.addTrack(command.getTrackId(), command.getLength());
             emit().onResponse(new ResponseEvents(game.getUncommittedChanges()));
         } else {
             emit().onError(new BusinessException(game.identity().value(),
