+++ src/main/java/co/com/sofkau/horseraces/usecases/game/commands/ChooseHorseUseCase.java
@@ -1,4 +1,4 @@
-package co.com.sofkau.horseraces.usecases;
+package co.com.sofkau.horseraces.usecases.game.commands;
 
 import co.com.sofka.business.generic.BusinessException;
 import co.com.sofka.business.generic.UseCase;
