+++ src/main/java/co/com/sofkau/horseraces/usecases/game/commands/DoRematchUseCase.java
@@ -1,27 +1,28 @@
-package co.com.sofkau.horseraces.usecases;
+package co.com.sofkau.horseraces.usecases.game.commands;
 
 import co.com.sofka.business.generic.BusinessException;
 import co.com.sofka.business.generic.UseCase;
 import co.com.sofka.business.support.RequestCommand;
 import co.com.sofka.business.support.ResponseEvents;
 import co.com.sofkau.horseraces.domain.game.Game;
-import co.com.sofkau.horseraces.domain.game.commands.RunRace;
+import co.com.sofkau.horseraces.domain.game.commands.DoRematch;
 
-public class RunRaceUseCase extends UseCase<RequestCommand<RunRace>, ResponseEvents> {
+public class DoRematchUseCase extends UseCase<RequestCommand<DoRematch>, ResponseEvents> {
+
 
     @Override
-    public void executeUseCase(RequestCommand<RunRace> runRaceRequestCommand) {
-        var command = runRaceRequestCommand.getCommand();
+    public void executeUseCase(RequestCommand<DoRematch> doRematchRequestCommand) {
+        var command = doRematchRequestCommand.getCommand();
         var game = Game.from(command.getGameId(), retrieveEvents(command.getGameId().value()));
 
-        if(game.getActualState().value().equals("PREPARED")){
-            game.runRace();
+        if (game.getActualState().value().equals("FINISHED")) {
+            game.doRematch();
             game.setPodium();
             emit().onResponse(new ResponseEvents(game.getUncommittedChanges()));
-        }
-        else {
+        } else {
             emit().onError(new BusinessException(game.identity().value(),
-                    "The game has not been prepared"));
+                    "The game has not finished"));
         }
     }
 }
+
