+++ src/main/java/co/com/sofkau/horseraces/usecases/game/commands/RunRaceUseCase.java
@@ -1,25 +1,27 @@
-package co.com.sofkau.horseraces.usecases;
+package co.com.sofkau.horseraces.usecases.game.commands;
 
 import co.com.sofka.business.generic.BusinessException;
 import co.com.sofka.business.generic.UseCase;
 import co.com.sofka.business.support.RequestCommand;
 import co.com.sofka.business.support.ResponseEvents;
 import co.com.sofkau.horseraces.domain.game.Game;
-import co.com.sofkau.horseraces.domain.game.commands.SetPodium;
+import co.com.sofkau.horseraces.domain.game.commands.RunRace;
 
-public class SetPodiumUseCase extends UseCase<RequestCommand<SetPodium>, ResponseEvents> {
+public class RunRaceUseCase extends UseCase<RequestCommand<RunRace>, ResponseEvents> {
+
     @Override
-    public void executeUseCase(RequestCommand<SetPodium> setPodiumRequestCommand) {
-        var command = setPodiumRequestCommand.getCommand();
+    public void executeUseCase(RequestCommand<RunRace> runRaceRequestCommand) {
+        var command = runRaceRequestCommand.getCommand();
         var game = Game.from(command.getGameId(), retrieveEvents(command.getGameId().value()));
 
-        if(game.getActualState().value().equals("RUNNING")){
+        if(game.getActualState().value().equals("PREPARED")){
+            game.runRace();
             game.setPodium();
             emit().onResponse(new ResponseEvents(game.getUncommittedChanges()));
         }
         else {
             emit().onError(new BusinessException(game.identity().value(),
-                    "The game actualState is not RUNNING"));
+                    "The game has not been prepared"));
         }
     }
 }
